+++ src/Models/Option/OptionDefinitions.cs
@@ -610,6 +610,19 @@ public static class OptionDefinitions
         }
     }
 
+    public static class Datadog
+    {
+        public const string DatadogResourceParam = "datadog-resource";
+
+        public static readonly Option<string> DatadogResourceName = new(
+            $"--{DatadogResourceParam}",
+            "The name of the Datadog resource to use. This is the unique name you chose for your Datadog resource in Azure."
+        )
+        {
+            IsRequired = true
+        };
+    }
+
     public static class KeyVault
     {
         public const string VaultNameParam = "vault";