+++ src/Program.cs
@@ -4,6 +4,7 @@
 using System.CommandLine.Builder;
 using AzureMcp.Commands;
 using AzureMcp.Services.Azure.AppConfig;
+using AzureMcp.Services.Azure.AzureIsv.Datadog;
 using AzureMcp.Services.Azure.Cosmos;
 using AzureMcp.Services.Azure.KeyVault;
 using AzureMcp.Services.Azure.Kusto;
@@ -96,6 +97,7 @@ internal class Program
         services.AddSingleton<ITenantService, TenantService>();
         services.AddSingleton<ICosmosService, CosmosService>();
         services.AddSingleton<IKustoService, KustoService>();
+        services.AddSingleton<IDatadogService, DatadogService>();
         services.AddSingleton<IStorageService, StorageService>();
         services.AddSingleton<IMonitorService, MonitorService>();
         services.AddSingleton<IMonitorHealthModelService, MonitorHealthModelService>();